web-preferences-api/reducedData.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>WebPreferences Test: reducedData attribute tests</title>
    <link rel="author" title="Luke Warlow" href="mailto:luke@warlow.dev" />
    <link rel="help" href="https://wicg.github.io/web-preferences-api/" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(() => {
        assert_true('preferences' in navigator, 'navigator.preferences defined');
        assert_true('reducedData' in navigator.preferences, 'navigator.preferences.reducedData defined');
      }, "navigator.preferences.reducedData supported");
      test(() => {
        navigator.preferences.reducedData = 'no-preference';
        assert_true(window.matchMedia('(prefers-reduced-data: no-preference)').matches, '(prefers-reduced-data: no-preference) matches');
        navigator.preferences.reducedData = 'reduce';
        assert_true(window.matchMedia('(prefers-reduced-data: reduce)').matches, '(prefers-reduced-data: reduce) matches');
        navigator.preferences.reducedData = null;
        assert_true(window.matchMedia('(prefers-reduced-data: no-preference)').matches, '(prefers-reduced-data: no-preference) matches');
      }, "navigator.preferences.reducedData updates (prefers-reduced-data)");
      test(() => {
        assert_throws_js(TypeError, () => { navigator.preferences.reducedData = 'this-is-an-invalid-value'; });
        assert_equals(null, navigator.preferences.reducedData);
        navigator.preferences.reducedData = null;
      }, "navigator.preferences.reducedData set to invalid value throws TypeError");
    </script>
  </body>
</html>
